% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{add_critique}
\alias{add_critique}
\title{Add criticism}
\usage{
add_critique(
  df_critique,
  Code,
  author,
  level,
  start_date,
  variable,
  type,
  comment = "",
  end_date = NULL,
  df_meta = NULL,
  resdir = NULL
)
}
\description{
Add criticism
}
