% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{apply_missing_year}
\alias{apply_missing_year}
\title{Long missing period over several years}
\usage{
apply_missing_year(Date, Value, Code, yearNA_lim)
}
\description{
Long missing period over several years
}
