% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_download}
\alias{CARD_download}
\title{CARD_download}
\usage{
CARD_download(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\link[base:logical]{logical} If TRUE, the current CARD version will be overwrite by the most recent one. Default, FALSE.}
}
\description{
Download the CARD repository in order to use last version of CARD available.
}
\seealso{
\itemize{
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
}
}
