% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{CARD_extraction}
\alias{CARD_extraction}
\title{CARD_extraction}
\usage{
CARD_extraction(
  data,
  CARD_path,
  WIP_dir = "WIP",
  period = NULL,
  samplePeriod_by_topic = NULL,
  simplify = TRUE,
  verbose = FALSE
)
}
\description{
CARD_extraction
}
