% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_list_all}
\alias{CARD_list_all}
\title{CARD_list_all}
\usage{
CARD_list_all()
}
\description{
List all the CARD variable.
}
\seealso{
\itemize{
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
}
}
