% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{CARD_transpose_data}
\alias{CARD_transpose_data}
\title{Transpose dataEX data.frame returned by CARD_extraction}
\usage{
CARD_transpose_data(dataEX)
}
\arguments{
\item{dataEX}{A \link{data.frame} returned by CARD_extraction with argument \code{simplify=TRUE}}
}
\value{
The transposed data frame with the first column as 'name' and following columns with the ID of the time series.
}
\description{
Transpose dataEX data.frame returned by CARD_extraction
}
