% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{CARD_trend}
\alias{CARD_trend}
\title{CARD_trend}
\usage{
CARD_trend(
  data,
  CARD_path,
  WIP_dir = "WIP",
  level = 0.1,
  period = NULL,
  samplePeriod_opti = NULL,
  simplify = TRUE,
  verbose = TRUE
)
}
\description{
CARD_trend
}
