% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_NApct}
\alias{compute_NApct}
\title{compute_NApct}
\usage{
compute_NApct(Value, Date, minDate, maxDate, sampleStart, sampleEnd)
}
\arguments{
\item{Value}{Numeric Vector of value on which the
percentage of NA is computed}

\item{minDate}{Date vector computed by 'process_extraction'}

\item{maxDate}{Date vector computed by 'process_extraction'}

\item{sampleStart}{Date vector computed by 'process_extraction'}

\item{sampleEnd}{Date vector computed by 'process_extraction'}
}
\value{
Percentage of NA
}
\description{
Compute the percentage of NA value in a sample.
}
