% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_extraction.R
\name{convert_dateEX}
\alias{convert_dateEX}
\title{convert_dateEX}
\usage{
convert_dateEX(data, isDate, nValue, verbose = TRUE)
}
\arguments{
\item{data}{A tibble containing a column named "Date" which corresponds to the index of each sample and one or several columns named "ValueEXx" (x being a number) that correspond to the extracted variables.}

\item{isDate}{A logical vector indicating whether the columns named "ValueEXx" should be converted to date format or not}

\item{nValue}{Number of columns named "ValueEXx"}

\item{verbose}{A logical, if TRUE output to the console will be provided}

\item{isColArgs}{A logical vector indicating whether the columns named "ValueEXx" corresponds to column arguments of function used in the extraction process}
}
\value{
A tibble containing a column named "Date" which corresponds to the date of each sample and one or several columns named "ValueEXx" (x being a number) that correspond to the extracted variables.
}
\description{
Convert the index to date format for the extracted variables specified in the isDate argument. The date format is based on the date column of the input tibble and the function "convert_data_hide" is used for the conversion.
}
\note{
documentation generated by chatGPT
}
