% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{copy_selection}
\alias{copy_selection}
\title{Copy a selection of station}
\usage{
copy_selection(
  select_dir,
  select_file,
  from_dir,
  to_dir,
  codeCol = "code",
  sep = ",",
  quote = "\\"",
  optname = "_HYDRO_QJM",
  codeSwipe = NULL,
  codeRm = NULL
)
}
\description{
Copy a selection of station
}
