% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{create_selection}
\alias{create_selection}
\title{Create a '.txt' selection of station}
\usage{
create_selection(computer_data_path, filedir, outname, optname = "_HYDRO_QJM")
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{filedir}{Directory of Banque HYDRO data you want to use in
ash\\computer_data_path\\ to get station codes. If "" is
use, data will be search in ash\\computer_data_path\\.}

\item{outname}{Name of the created selection file.}

\item{optname}{Optional name that will be added to the selected
station codes to match names of the Banque HYDRO files
(default: '_HYDRO_QJM').}
}
\value{
Writes a selection '.txt' file of data filename of
stations.
}
\description{
Create a txt file that resume all the station data
files present in a filedir
}
