% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{datasheet_panel}
\alias{datasheet_panel}
\title{Datasheet panel}
\usage{
datasheet_panel(
  list_df2plot,
  df_meta,
  trend_period,
  mean_period,
  linetype_per,
  axis_xlim,
  colorForce,
  info_header,
  time_header,
  foot_note,
  structure,
  layout_matrix,
  info_height,
  time_height,
  var_ratio,
  foot_height,
  paper_size,
  df_shapefile,
  logo_path,
  zone_to_show,
  show_colorEvent,
  outdirTmp_pdf,
  outdirTmp_png,
  df_page = NULL
)
}
\description{
Datasheet panel
}
