% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{estimateHurst}
\alias{estimateHurst}
\title{Hurst coefficient}
\usage{
estimateHurst(Z, DoDetrending = TRUE, trend = getMKStat(Z)$trend)
}
\arguments{
\item{Z}{numeric, data (NA-free)}

\item{DoDetrending}{logical, detrend data before estimating Hurst?}

\item{trend}{numeric, trend value (only used if detrending
required)}
}
\value{
The estimated value of the hurst coefficient
}
\description{
Estimate the Hurst coefficient of a series
}
\examples{
\dontrun{
data(nhtemp) #Average Yearly Temperatures in New Haven
estimateHurst(Z=nhtemp)
}
}
