% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXstat.R
\name{EXstat}
\alias{EXstat}
\title{EXstat}
\description{
Provides an efficient and simple solution to aggregate and analyze the stationarity of time series.
}
\details{
See \href{https://github.com/super-lou/EXstat}{EXstat github documentation} \url{https://github.com/super-lou/EXstat} for an easier approach.

For a more user-friendly interaction, this package has been developed in symbiosis with predefined parameter files called \href{https://github.com/super-lou/CARD/}{CARD} \url{https://github.com/super-lou/CARD/}.
So you don't have to define complex parameters yourself to extract hydroclimatological variables. What's more, if the CARD you want doesn't exist, it's easy to create one based on the others.
}
\seealso{
\itemize{
\item \code{\link[=process_extraction]{process_extraction()}} for extracting variables.
\item \code{\link[=process_trend]{process_trend()}} for performing trend analysis on extracted variables.
\item \code{\link[=CARD_extraction]{CARD_extraction()}} for extracting variables using CARD parameterization files.
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
}
}
\author{
Main Developer and Maintainer
\itemize{
\item Louis Héraut \href{mailto:louis.heraut@inrae.fr}{louis.heraut@inrae.fr} INRAE
}

Other Developer
\itemize{
\item Benjamin Renard \href{mailto:benjamin.renard@inrae.fr}{benjamin.renard@inrae.fr} INRAE
\item Valentin Mansanarez
}

Other Contributors
\itemize{
\item Éric Sauquet \href{mailto:eric.sauquet@inrae.fr}{eric.sauquet@inrae.fr} INRAE
\item Jean-Philippe Vidal \href{mailto:jean-philippe.vidal@inrae.fr}{jean-philippe.vidal@inrae.fr} INRAE
\item Nathan Pellerin
}
}
