% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_climate_data}
\alias{extract_climate_data}
\title{Extract climate data}
\usage{
extract_climate_data(
  computer_data_path,
  filedir,
  colNames = c("Date", "PRCP_mm", "PET_mm", "T_degC"),
  verbose = TRUE
)
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{filedir}{Directory of data you want to use in
ash\\computer_data_path\\ to get climate data. If "" is
use, data will be search in ash\\computer_data_path\\.}

\item{colNames}{String or vector of string of column names of the
tibble that will be created.}

\item{verbose}{Boolean to indicate if more processing info are
printed (default : TRUE).}
}
\value{
A tibble containing data about selected stations.
}
\description{
Extraction of climate data and metadata
}
