% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_data}
\alias{extract_data}
\title{Extract data}
\usage{
extract_data(computer_data_path, filedir, filename, verbose = TRUE)
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{filedir}{Directory of Banque HYDRO data you want to use in
ash\\computer_data_path\\ to get station codes. If "" is
use, data will be search in ash\\computer_data_path\\.}

\item{filename}{String or vector of string of all filenames from
which data will be extracted. If set to 'all', all the file in
'fildir' will be use.}

\item{verbose}{Boolean to indicate if more processing info are
printed (default : TRUE).}
}
\value{
A tibble containing data about selected stations.
}
\description{
Extraction of data of stations.
}
