% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{extract_Var_WRAP}
\alias{extract_Var_WRAP}
\title{extract_Var WRAP}
\usage{
extract_Var_WRAP(
  df_data,
  funct,
  period,
  hydroPeriod = "01-01",
  timestep = "year",
  isDate = FALSE,
  ...
)
}
\description{
extract_Var WRAP
}
