% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_process.R
\name{extraction_process}
\alias{extraction_process}
\title{extraction_process}
\usage{
extraction_process(
  data,
  funct = max,
  funct_args = list(),
  timeStep = "year",
  samplePeriod = NULL,
  period = NULL,
  isDate = FALSE,
  NApct_lim = 10,
  NAyear_lim = 10,
  Seasons = c("DJF", "MAM", "JJA", "SON"),
  onlyDate4Season = FALSE,
  nameEX = "X",
  keep = FALSE,
  compress = FALSE,
  rmNApct = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Input data format is a tibble of at least a column of
date, a column of numeric value and a character column for names of
time series in order to identify them. Thus it is possible to have
a tibble with multiple time series which can be grouped by their
names.}

\item{funct}{The function (for example \code{mean}) that you want to
use. Arguments of the chosen function can be passed to this
extraction process and the function can be previously defined.}

\item{timeStep}{A character chain which can be "year" for yearly
extraction and "year-month" for monthly extraction along years.}

\item{samplePeriod}{A vector of two character chains to precise the
sampling of the extraction (for example, in a yearly extraction,
c("05-01", "11-30") will use only the data from the 1st of may to
the 30th of november). It can also just be a simple character chain
(as "02-01" in a yearly extraction) to start the sampling the 1st
of february and to end it the 31th of january (hence it is similar
to c("02-01", "01-31")).}

\item{period}{A vector of two date to restrict the period of
analysis. Imposed date format is "YYYY-mm-dd". As an example it can
be c("1950-01-01", "2020-12-31") to select data from the 1st
january of 1950 to the end of december of 2020. The default option
is period=NULL which considers all available data.}

\item{...}{Arguments needed for the function provided through
\code{funct}.}
}
\value{
The output is also a tibble with a column of date, of
character for the name of time series and a numerical column with
the extracted variable from the time series.
}
\description{
This process extract a variable from time series (for
example the yearly mean of time series). Extraction can be yearly,
monthly or realised each month of every year and sampled
differently along years or months.
}
\examples{
# Date
Start = as.Date("1972-01-01")
End = as.Date("2020-12-31")
Date = seq.Date(Start, End, by="day")

# Value to analyse
set.seed(100)
X = seq(1, length(Date))/1e4 + runif(length(Date), -100, 100)
X[as.Date("2000-03-01") <= Date & Date <= as.Date("2000-09-30")] = NA

# Creation of tibble
data = dplyr::tibble(Date=Date, ID="serie A", X=X)

# Extraction
dataEX = extraction_process(data=data,
                            samplePeriod=c("05-01",
                                           "11-30"),
                            funct=max,
                            na.rm=TRUE,
                            period=c(as.Date("1990-01-01"),
                                     as.Date("2020-12-31")),
                            timeStep="year")
dataEX
}
