% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fieldSignificance_FDR}
\alias{fieldSignificance_FDR}
\title{FDR field significance}
\usage{
fieldSignificance_FDR(pvals, level = 0.1)
}
\arguments{
\item{pvals}{numeric vector, p-values of local tests}

\item{level}{numeric in (0, 1), level at which field significance
is evaluated}
}
\value{
pFDR, the FDR p-value, interpreted as follows: local
p-values smaller than pFDR are field-significant.
}
\description{
Field significance using the false detection rate
approach
}
\examples{
set.seed(123456) # Make example reproducible
level = 0.1 # Level of the test
data(nhtemp) # Average Yearly Temperatures in New Haven
# Add 29 stationary series to nhtemp series
X = matrix(c(nhtemp, rnorm(length(nhtemp)*29)),
             nrow=length(nhtemp))
# Compute local p-values from a MK test
pvals = rep(NA, 30)
for (i in 1:30) {
    pvals[i] = generalMannKendall(X[, i], level=level)$P
}
# Evaluate field significance
pFDR = fieldSignificance_FDR(pvals, level)
which(pvals <= level) # Locally-significant sites
which(pvals <= pFDR) # FDR-significant sites
}
\references{
Benjamini, Y., and Y. Hochberg (1995), Controlling the
false discovery rate: A practical and powerful approach to multiple
testing, J. R. Stat. Soc., Ser. B., 57, 289–300.
}
