% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expostat.R
\name{FuncExit}
\alias{FuncExit}
\title{End the analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStat.}
}
\value{
Exit status
}
\description{
End the analysis
}
\examples{
res = InitStat()
   res1 = LoadStat(PID = res$PID, UseExample = "example#1")
   res2 = StatCorr(PID = res$PID, Group = T, VarsX = "X5,X6,X7,X8,X9", VarsY = "Y1", VarsBy = "Y1", Method = "pearson", Layout = "bubble", Brightness = "dark", Palette = "nature")
   res3 = FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
