% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{GeneralMannKendall_WRAP}
\alias{GeneralMannKendall_WRAP}
\title{Mann-Kendall trend analysis  commit 1}
\usage{
GeneralMannKendall_WRAP(
  X,
  level = 0.1,
  timeDep_option = "INDE",
  DoDetrending = TRUE
)
}
\arguments{
\item{X}{data (vector). IMPORTANT: it assumes that X is
regularly-spaced.}

\item{level}{numeric, between 0 and 1, level of the test (default :
0.1).}

\item{DoDetrending}{logical, only used for when dep.option == 'LTP'
(default: TRUE).}

\item{option}{dependency option, option for handling temporal
dependence
(default 'INDE' for independence).}
}
\value{
a dataframe, with the different statistics and values of
interest of the test.
}
\description{
Apply the generalMannKendall function to the serie X.
}
\examples{
GeneralMannKendall_WRAP(X=1:100)
GeneralMannKendall_WRAP(X=rep(1, 100))
}
