% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{get_CARD_metadata}
\alias{get_CARD_metadata}
\title{get_CARD_metadata}
\usage{
get_CARD_metadata(CARD_path)
}
\arguments{
\item{CARD_path}{A \link[base:character]{character} string representing the path to the downloaded CARD directory (it should end with \code{"CARD"}). In this directory, you can copy and paste (and later modify) CARDs from the \code{"__all__"} subdirectory that you want to use for an analysis represented by a subdirectory named \code{CARD_dir} (see \code{CARD_tmp} if you want to locate your \code{CARD_dir} directory elsewhere). In your CARDs, you can specify functions available in the scripts of the \code{"__tools__"} subdirectory.}
}
\description{
Get all the CARD metadata in a \link[tibble:tibble]{tibble}.
}
