% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trend.R
\name{get_intercept}
\alias{get_intercept}
\title{Intercept of trend}
\usage{
get_intercept(dataEX, trendEX, verbose = TRUE)
}
\arguments{
\item{dataEX}{DataEX frame that contains at least columns for "Code","Date" and "Value"}

\item{trendEX}{DataEXframe that contains the result of trend analysis}

\item{verbose}{logical value. If set to True, it print the result of the computation.}
}
\value{
A dataframe which contains the input dataframe trendEX with an additional column 'b' that represents the intercept of the trend.
}
\description{
Computes the intercept values of linear trends using the first order values of trends and the dataEX on which the analysis is performed.
}
\note{
documentation generated by chatGPT
}
