% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trend.R
\name{get_period}
\alias{get_period}
\title{get_period}
\usage{
get_period(dataEX, trendEX, verbose = TRUE)
}
\arguments{
\item{dataEX}{A data frame containing the data.}

\item{trendEX}{A data frame containing the results of the trend analysis.}

\item{verbose}{A logical value indicating whether to display verbose output (default is TRUE).}
}
\value{
A modified data frame with an additional column indicating the optimal periods for trend analysis.
}
\description{
Computes the optimal periods for trend analysis based on the data.
}
\details{
The function computes the optimal periods for trend analysis based on the data. It calculates the minimum and maximum dates (start and end) for each code group in the data. Then, it creates a new column called "period" that contains a list of the start and end dates for each code group. The function then joins this information with the trend analysis results using the "Code" column.
}
\note{
documentation generated by chatGPT
}
