% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_selection_DOCX}
\alias{get_selection_DOCX}
\title{Get selection of a '.docx' file}
\usage{
get_selection_DOCX(
  computer_data_path,
  listdir,
  listname,
  code_nameCol = "code",
  choice_nameCol = "Choix",
  choice_Val = c("A garder", "Ajout"),
  optname = "_HYDRO_QJM"
)
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{listdir}{Directory of the selection '.docx' file in
ash\\computer_data_path\\.}

\item{listname}{Name of the selection '.docx' file.}

\item{code_nameCol}{Name of the column of the file containing code
stations}

\item{choice_nameCol}{Name of the column of the file containing the
choice if the associated station is selected or not. If set to
'NULL', all stations in the 'code_nameCol' are selected.}

\item{choice_Val}{Vector of string meaning that the station is
selected.}

\item{optname}{Optional name that will be added to the selected
station codes to match names of the Banque HYDRO files
(default: '_HYDRO_QJM').}
}
\value{
A vector of string containing file names of stations.
}
\description{
Gets the selection of station from a formated '.docx'
file.
}
