% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_selection_TXT}
\alias{get_selection_TXT}
\title{Get selection of a '.txt' file}
\usage{
get_selection_TXT(computer_data_path, listdir, listname)
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{listdir}{Directory of the selection '.txt' file in
ash\\computer_data_path\\.}

\item{listname}{Name of the selection '.txt' file.}
}
\value{
A vector of string containing file names of stations.
}
\description{
Gets the selection of station from the selection txt
file generated by the 'create_selection' function
}
