% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R, R/analyse_save.R
\name{get_Xtrend}
\alias{get_Xtrend}
\title{X trend}
\usage{
get_Xtrend(
  var,
  df_data,
  period,
  level = 0.1,
  df_flag = NULL,
  NApct_lim = NULL,
  NAyear_lim = NULL,
  day_to_roll = NULL,
  functM = NULL,
  functM_args = NULL,
  isDateM = FALSE,
  samplePeriodM = NULL,
  isAlongYearM = TRUE,
  functS = NULL,
  functS_args = NULL,
  isDateS = FALSE,
  samplePeriodS = NULL,
  isAlongYearS = TRUE,
  functY = NULL,
  functY_args = NULL,
  isDateY = FALSE,
  samplePeriodY = NULL,
  functYT_ext = NULL,
  functYT_ext_args = NULL,
  isDateYT_ext = FALSE,
  functYT_sum = NULL,
  functYT_sum_args = NULL,
  functG = NULL,
  functG_args = NULL,
  df_mod = tibble(),
  verbose = TRUE
)

get_Xtrend(
  var,
  df_data,
  period,
  level = 0.1,
  df_flag = NULL,
  NApct_lim = NULL,
  NAyear_lim = NULL,
  day_to_roll = NULL,
  functM = NULL,
  functM_args = NULL,
  isDateM = FALSE,
  samplePeriodM = NULL,
  isAlongYearM = TRUE,
  functS = NULL,
  functS_args = NULL,
  isDateS = FALSE,
  samplePeriodS = NULL,
  isAlongYearS = TRUE,
  functY = NULL,
  functY_args = NULL,
  isDateY = FALSE,
  samplePeriodY = NULL,
  functYT_ext = NULL,
  functYT_ext_args = NULL,
  isDateYT_ext = FALSE,
  functYT_sum = NULL,
  functYT_sum_args = NULL,
  functG = NULL,
  functG_args = NULL,
  df_mod = tibble(),
  verbose = TRUE
)
}
\description{
X trend

X trend
}
