% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getMKStat}
\alias{getMKStat}
\title{Mann-Kendall statistics}
\usage{
getMKStat(X)
}
\arguments{
\item{X}{numeric, data}
}
\value{
A list, with components: $stat, MK statistics; $trend,
Sen's estimate.
}
\description{
Compute MK stat and Sen's trend estimate
}
\examples{
\dontrun{
data(nhtemp) #Average Yearly Temperatures in New Haven
getMKStat(X = nhtemp)
}
}
