% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getTiesCorrection}
\alias{getTiesCorrection}
\title{Ties correction}
\usage{
getTiesCorrection(Z)
}
\arguments{
\item{Z}{numeric, data (NA-free)}
}
\value{
the correction for the variance of MK stat
}
\description{
Compute correction to the variance of MK statistics to
account for ties
}
\examples{
\dontrun{
data(nhtemp) # Average Yearly Temperatures in New Haven
getTiesCorrection(Z=nhtemp)
getTiesCorrection(rnorm(100))
}
}
