% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{HurstLkh}
\alias{HurstLkh}
\title{Hurst likelihood}
\usage{
HurstLkh(H, x)
}
\arguments{
\item{H}{numeric,  hurst coeff. value}

\item{x, }{data sample}
}
\value{
log-likelihood value
}
\description{
Compute the likelihood function to be maximized for
estimating the Hurst coefficient H
}
\examples{
\dontrun{
data(nhtemp) # Average Yearly Temperatures in New Haven
HurstLkh(H=0.8, x=nhtemp)
HurstLkh(H=0.5, x=nhtemp)
HurstLkh(H=0.2, x=nhtemp)
}
}
