% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{info_panel}
\alias{info_panel}
\title{Info panel}
\usage{
info_panel(
  list_df2plot,
  df_meta,
  trend_period = NULL,
  mean_period = NULL,
  period = NULL,
  df_shapefile = NULL,
  codeLight = NULL,
  df_data_code = NULL,
  to_do = "all",
  zone_to_show = "France"
)
}
\description{
Info panel
}
