% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_panel}
\alias{layout_panel}
\title{Layout panel}
\usage{
layout_panel(
  df_data,
  df_meta,
  structure,
  layout_matrix,
  to_plot = c("datasheet", "table", "map", "map_regime", "map_trend", "map_mean"),
  figdir = "",
  filedir_opt = "",
  filename_opt = "",
  variable = "",
  df_trend = NULL,
  alpha = 0.1,
  unit2day = 365.25,
  var = "",
  type = "",
  event = "",
  unit = "",
  glose = NULL,
  trend_period = NULL,
  mean_period = NULL,
  colorForce = FALSE,
  exQprob = 0.01,
  linetype_per = "solid",
  axis_xlim = NULL,
  missRect = TRUE,
  paper_size = "A4",
  time_header = NULL,
  info_header = NULL,
  foot_note = TRUE,
  info_height = 2.8,
  time_height = 3,
  var_ratio = 3,
  foot_height = 1.25,
  df_shapefile = NULL,
  resdir = NULL,
  logo_path = NULL,
  zone_to_show = NULL,
  pdf_chunk = c("all"),
  show_colorEvent = FALSE
)
}
\description{
Layout panel
}
