% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{load_logo}
\alias{load_logo}
\title{Logo loading}
\usage{
load_logo(
  resources_path,
  logo_dir,
  PRlogo_file,
  AEAGlogo_file,
  INRAElogo_file,
  FRlogo_file,
  logo_to_show
)
}
\description{
Logo loading
}
