% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{load_shapefile}
\alias{load_shapefile}
\title{Shapefiles loading}
\usage{
load_shapefile(
  resources_path,
  df_meta,
  fr_shpdir,
  fr_shpname,
  bs_shpdir,
  bs_shpname,
  sbs_shpdir,
  sbs_shpname,
  cbs_shpdir,
  cbs_shpname,
  cbs_coord,
  rv_shpdir,
  rv_shpname,
  river_selection = c("all")
)
}
\arguments{
\item{resources_path}{Path to the resources directory.}

\item{fr_shpdir}{Directory you want to use in ash\\resources_path\\
to get the France shapefile.}

\item{fr_shpname}{Name of the France shapefile.}

\item{bs_shpdir}{Directory you want to use in ash\\resources_path\\
to get the hydrological basin shapefile.}

\item{bs_shpname}{Name of the hydrological basin shapefile.}

\item{sbs_shpdir}{Directory you want to use in
ash\\resources_path\\ to get the hydrological sub-basin shapefile.}

\item{sbs_shpname}{Name of the hydrological sub-basin shapefile.}

\item{rv_shpdir}{Directory you want to use in ash\\resources_path\\
to get the hydrological network shapefile.}

\item{rv_shpname}{Name of the hydrological network shapefile.}

\item{show_river}{Boolean to indicate if the shapefile of the
hydrological network will be charge because it is a heavy one and
that it slows down the entire process (default : TRUE)}
}
\value{
A list of shapefiles converted as tibbles that can be plot
with 'geom_polygon' or 'geom_path'.
}
\description{
Generates a list of shapefiles to draw a hydrological
map of the France
}
