% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_panel}
\alias{map_panel}
\title{Map panel}
\usage{
map_panel(
  list_df2plot,
  df_meta,
  df_shapefile,
  idPer_trend = 1,
  trend_period,
  mean_period,
  colorForce = FALSE,
  exQprob = 0.01,
  codeLight = NULL,
  mapType = "trend",
  margin = NULL,
  showSea = TRUE,
  foot_note = FALSE,
  foot_height = 0,
  logo_path = NULL,
  zone_to_show = "France",
  df_page = NULL,
  outdirTmp_pdf = "",
  outdirTmp_png = "",
  verbose = TRUE
)
}
\description{
Map panel
}
