% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_extraction.R
\name{missing_year}
\alias{missing_year}
\title{Missing year}
\usage{
missing_year(data, nValue, NAyear_lim = 10, mod = NULL, verbose = TRUE)
}
\arguments{
\item{data}{A data.frame containing at least a date column, a numeric column and a column indicating a name of time series, in order to identify them.}

\item{nValue}{number of numeric column}

\item{NAyear_lim}{threshold of year duration for missing period.}

\item{mod}{An object of class modification, it will store all modification made to the data}

\item{verbose}{logical for print prints}
}
\value{
return the dataframe with identified missing period as NA
}
\description{
This function will check for any continuous missing period longer than a specified threshold of years in each time series and assign them as missing values.
}
\note{
documentation generated by chatGPT

documentation generated by chatGPT
}
