% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_process.R
\name{missing_year_hide}
\alias{missing_year_hide}
\title{Long missing period over several years}
\usage{
missing_year_hide(Value, Date, NAyear_lim)
}
\description{
Long missing period over several years
}
