% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_extraction.R
\name{NA_filter}
\alias{NA_filter}
\title{NA filter}
\usage{
NA_filter(data, timeStep, nValue, NApct_lim = 1, mod = NULL, verbose = TRUE)
}
\arguments{
\item{data}{Data after extraction that should be filtered.}

\item{timeStep}{Character string indicating the time step of the extraction, could be 'none' if the time step doesn't matter, 'year' or 'year-month'.}

\item{nValue}{Number of values extracted.}

\item{NApct_lim}{Percentage limit of NA above which rows will be removed. Default is 1.}

\item{mod}{A list of modification to store comments of filtering applied.}

\item{verbose}{Logical, when true prints print indicating the removal of data due to high NA percentage}
}
\value{
data filtered by the percentage of NA allowed in input.
}
\description{
Filters the rows with a NA percentage above the threshold given in parameter.
}
\note{
documentation generated by chatGPT
}
