% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trend.R
\name{process_trend}
\alias{process_trend}
\title{process_trend}
\usage{
process_trend(
  dataEX,
  metaEX = NULL,
  MK_level = 0.1,
  time_dependency_option = "INDE",
  suffix = NULL,
  suffix_delimiter = "_",
  extreme_take_not_signif_into_account = TRUE,
  extreme_take_only_id = NULL,
  extreme_by_suffix = TRUE,
  period_trend = NULL,
  period_change = NULL,
  extreme_prob = 0.01,
  show_advance_stat = FALSE,
  dev = FALSE,
  verbose = FALSE,
  verbose_stat = FALSE
)
}
\arguments{
\item{dataEX}{A data frame containing the data to be analyzed.}

\item{metaEX}{A data frame containing metadata information (optional).}

\item{MK_level}{The significance level for the Mann-Kendall statistical test.}

\item{time_dependency_option}{The time dependency option to use in the Mann-Kendall test.}

\item{extreme_take_not_signif_into_account}{Flag indicating whether to take into account non-significant trends.}

\item{period_trend}{A list of periods to consider for trend analysis (optional).}

\item{period_change}{A list of periods to consider for change analysis (optional).}

\item{extreme_prob}{The extreme probability threshold for identifying extreme values (optional).}

\item{verbose}{Flag indicating whether to display verbose output during the analysis.}

\item{...}{Additional parameters to be passed to the statistical test and estimation methods.}
}
\value{
A data frame containing the results of the trend analysis.
}
\description{
Analyzes the trend in data using statistical tests and estimation methods.
}
\details{
The function analyzes the trend in the provided data using statistical tests, such as the Mann-Kendall test, and estimation methods. It computes the trend for each specified period, detects significant trends, estimates other variables related to the trend, and identifies extreme values if metadata information is provided.
}
\note{
documentation generated by chatGPT
}
\examples{
# Date
Start = as.Date("1972-01-01")
End = as.Date("2020-12-31")
Date = seq.Date(Start, End, by="day")

# Variable to analyse
set.seed(100)
X = seq(1, length(Date))/1e4 + runif(length(Date), -100, 100)
X[as.Date("2000-03-01") <= Date & Date <= as.Date("2000-09-30")] = NA

# Creation of tibble
data = dplyr::tibble(date=Date, ID="serie A", X=X)

# Extraction
dataEX = process_extraction(data=data,
                            sampling_period=c("05-01",
                                           "11-30"),
                            funct=max,
                            na.rm=TRUE,
                            period=c(as.Date("1990-01-01"),
                                     as.Date("2020-12-31")),
                            time_step="year")

trendEX = process_trend(data=dataEX)
trendEX

}
