% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{randomizedNormalScore}
\alias{randomizedNormalScore}
\title{Randomized Normal Score}
\usage{
randomizedNormalScore(x)
}
\arguments{
\item{x}{numeric, data}
}
\value{
the normal-score-transformed series
}
\description{
Randomized Normal Score transformation
}
\examples{
\dontrun{
data(nhtemp) # Average Yearly Temperatures in New Haven
z = randomizedNormalScore(x=nhtemp)
par(mfrow = c(1, 2))
plot(nhtemp, type='b');plot(z, type='b')
}
}
