% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R, R/analyse_save.R
\name{rollmean_code}
\alias{rollmean_code}
\title{Rolling average station}
\usage{
rollmean_code(df_data, nroll = 10, df_mod = NULL, verbose = TRUE)

rollmean_code(df_data, nroll = 10, df_mod = NULL, verbose = TRUE)
}
\description{
Rolling average station

Rolling average station
}
