% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{sampling_data}
\alias{sampling_data}
\title{Sampling data}
\usage{
sampling_data(
  df_data,
  df_meta,
  hydroPeriod = c("05-01", "11-30"),
  Code = NULL,
  df_mod = NULL
)
}
\description{
Sampling data
}
