% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{short_trendExtremes}
\alias{short_trendExtremes}
\title{Trend Extremes}
\usage{
short_trendExtremes(
  list_df2plot,
  Code,
  nPeriod_trend,
  nbVar,
  nCode,
  colorForce = FALSE
)
}
\description{
Trend Extremes
}
