% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{sourceProcess}
\alias{sourceProcess}
\title{sourceProcess}
\usage{
sourceProcess(path, default = NULL)
}
\arguments{
\item{path}{directory path to the R script file.}

\item{default}{a default process list to merge with the process read from the script file.}
}
\value{
a list of processes defined in the R script file.
}
\description{
The function reads an R script file and extract the defined process.
}
\note{
documentation generated by chatGPT
}
