% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_panel}
\alias{table_panel}
\title{Table panel}
\usage{
table_panel(
  list_df2plot,
  df_meta,
  trend_period,
  mean_period,
  colorForce = FALSE,
  exQprob = 0.01,
  slice = NULL,
  title = NULL,
  paper_size = "A3",
  foot_note = FALSE,
  foot_height = 0,
  resdir = NULL,
  logo_path = NULL,
  outdirTmp_pdf = "",
  outdirTmp_png = "",
  df_page = NULL
)
}
\description{
Table panel
}
