% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{time_panel}
\alias{time_panel}
\title{Time panel}
\usage{
time_panel(
  df_data_code,
  df_trend_code,
  var,
  type,
  unit,
  hydroPeriod_code = NULL,
  linetype_per = "solid",
  alpha = 0.1,
  colorForce = FALSE,
  missRect = FALSE,
  unit2day = 365.25,
  trend_period = NULL,
  mean_period = NULL,
  axis_xlim = NULL,
  grid = TRUE,
  ymin_lim = NULL,
  color = NULL,
  NspaceMax = NULL,
  first = FALSE,
  last = FALSE,
  lim_pct = 10
)
}
\description{
Time panel
}
