% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_extraction.R
\name{card_extraction}
\alias{card_extraction}
\title{card_extraction}
\usage{
card_extraction(
  data,
  card_dir,
  period = NULL,
  samplePeriod_opti = NULL,
  verbose = TRUE
)
}
\description{
card_extraction
}
