% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{CARD_extraction}
\alias{CARD_extraction}
\title{CARD_extraction}
\usage{
CARD_extraction(
  data,
  CARD_path,
  CARD_dir = "WIP",
  CARD_name = NULL,
  CARD_tmp = NULL,
  period_default = NULL,
  suffix = NULL,
  cancel_lim = FALSE,
  simplify = FALSE,
  expand_overwrite = NULL,
  sampling_period_overwrite = NULL,
  rmNApct = TRUE,
  rm_duplicates = FALSE,
  dev = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Tibble used in the extraction}

\item{CARD_path}{Directory path where CARD_dir is located}

\item{CARD_dir}{Subdirectory name where the script files are located (default = "WIP")}

\item{CARD_name}{Name of the script files to be used (optional)}

\item{CARD_tmp}{Temporary directory path (optional)}

\item{period_default}{Period to extract from the data (optional)}

\item{suffix}{Suffix to append to extracted variables (optional)}

\item{cancel_lim}{Specify whether to cancel limits (default = FALSE)}

\item{simplify}{Specify whether to simplify the extracted data by column name (default = FALSE)}

\item{expand_overwrite}{Overwrite the expand parameter (optional)}

\item{sampling_period_overwrite}{Overwrite the sampling_period parameter (optional)}

\item{verbose}{Specify whether to print out the process details (default = FALSE)}
}
\value{
A list of extracted data, along with meta data.
}
\description{
Extract specified data from a set of script files.
}
\details{
The function reads the script files from the specified directory and extracts data based on the parameters provided.
}
