% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{CARD_extraction}
\alias{CARD_extraction}
\title{CARD_extraction}
\usage{
CARD_extraction(
  data,
  CARD_path,
  CARD_dir = "WIP",
  CARD_name = NULL,
  CARD_tmp = NULL,
  period = NULL,
  suffix = NULL,
  cancel_lim = FALSE,
  simplify = NULL,
  expand_overwrite = NULL,
  samplePeriod_overwrite = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{Tibble used in the extraction}

\item{CARD_path}{directory path where CARD_dir is located}

\item{CARD_dir}{default="WIP", specify the name of the subdirectory in which to find the script files.}

\item{CARD_name}{specify the name of the script files to be used}

\item{period}{specify a period to extract from data}

\item{simplify}{specify whether or not to simplify the extracted data by a column name}

\item{verbose}{default=FALSE, if set to TRUE, the function will print out the process details}
}
\value{
a list of extracted data, along with meta data.
}
\description{
The function extract specified data from a set of script files.
}
\note{
documentation generated by chatGPT
}
