% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_list_all}
\alias{CARD_list_all}
\title{CARD_list_all}
\usage{
CARD_list_all()
}
\description{
List all the CARD parametrization files in a \link[tibble:tibble]{tibble}.
}
\seealso{
\enumerate{
\item \code{\link[=CARD_download]{CARD_download()}} for downloading last version of CARD parameterization files.
}
\itemize{
\item 
\enumerate{
\item \code{\link[=CARD_list_all]{CARD_list_all()}} list all available CARD.
}
}
\enumerate{
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
\item \code{\link[=CARD_extraction]{CARD_extraction()}} for extracting variables using CARD.
\item \code{\link[=process_extraction]{process_extraction()}} for extracting variables.
\item \code{\link[=process_trend]{process_trend()}} for performing trend analysis on extracted variables.
}
}
