% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_management}
\alias{CARD_management}
\title{CARD_management}
\usage{
CARD_management(
  CARD = ".",
  tmp = "",
  layout = c("EX", "[", "QA", "]"),
  underscore_to_white = TRUE,
  add_id = TRUE,
  overwrite = TRUE,
  verbose = FALSE,
  args = NULL
)
}
\arguments{
\item{CARD}{The CARD directory path (need to ends by "CARD").}

\item{tmp}{The temporary directory path. If not provided, it will default to the value of \code{CARD}.}

\item{layout}{A character vector specifying the layout of files to be managed. The default layout is c("EX", "\link{", "QA", "}").}

\item{underscore_to_white}{Logical. If TRUE, underscores in file names will be replaced with spaces. Default is TRUE.}

\item{add_id}{Logical. If TRUE, numerical id will be added to the start of the file names. Default is TRUE.}

\item{overwrite}{Logical. If TRUE, existing files in the temporary directory will be overwritten. Default is TRUE.}

\item{verbose}{Logical. If TRUE, intermediate prints will be printed during the execution of the function. Default is FALSE.}

\item{args}{A list of arguments needed for the function. If not provided, it will be automatically created using the other function arguments. It is an intermediate form of arguments useful if argparse package is used.}
}
\description{
This function manages the CARD directory structure and performs file operations based on the specified arguments.
}
\examples{
# Manage CARD directory with default arguments
CARD_management(CARD = "/path/to/CARD")

# Manage CARD directory with custom arguments
CARD_management(CARD="/path/to/CARD", tmp="/path/to/tmp", layout=c("EX", "[", "QA", "]"),
                underscore_to_white=TRUE, add_id=TRUE, overwrite=TRUE, verbose=TRUE)

}
