% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_management}
\alias{CARD_management}
\title{CARD_management}
\usage{
CARD_management(
  CARD = ".",
  tmp = "",
  layout = c("EX", "[", "QA", "]"),
  white = TRUE,
  blank = FALSE,
  overwrite = TRUE,
  verbose = FALSE,
  args = NULL
)
}
\description{
A
}
