% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{CARD_trend}
\alias{CARD_trend}
\title{CARD_trend}
\usage{
CARD_trend(
  data,
  CARD_path,
  CARD_dir = "WIP",
  CARD_name = NULL,
  simplify_by = NULL,
  level = 0.1,
  period = NULL,
  samplePeriod_by_topic = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{Tibble used in the extraction}

\item{CARD_path}{directory path where CARD_dir is located}

\item{CARD_dir}{default="WIP", specify the name of the subdirectory in which to find the script files.}

\item{CARD_name}{specify the name of the script files to be used}

\item{simplify_by}{specify whether or not to simplify the extracted data}

\item{level}{default=0.1, level for the trend analysis}

\item{period}{specify a period to extract from data}

\item{samplePeriod_by_topic}{specify a sample period per topic}

\item{verbose}{default=TRUE, if set to TRUE, the function will print out the process details}
}
\value{
a list of extracted data, meta data, and trend data.
}
\description{
The function applies trend analysis on the extracted data. It uses the CARD_extraction function to extract data first.
}
\note{
documentation generated by chatGPT
}
