% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXstat.R
\name{EXstat}
\alias{EXstat}
\title{EXstat}
\description{
Provides an efficient and simple solution to aggregate and analyze the stationarity of time series.
}
\details{
See \href{https://github.com/super-lou/EXstat}{EXstat github documentation} \url{https://github.com/super-lou/EXstat} for an easier approach.

For a more user-friendly interaction, this package has been developed in symbiosis with predefined parameter files called \href{https://github.com/super-lou/CARD/}{CARD} \url{https://github.com/super-lou/CARD/}.
So you don't have to define complex parameters yourself to extract hydroclimatological variables. What's more, if the CARD you want doesn't exist, it's easy to create one based on the others.
}
\seealso{
\enumerate{
\item \code{\link[=process_extraction]{process_extraction()}} for extracting variables.
\item \code{\link[=process_trend]{process_trend()}} for performing trend analysis on extracted variables.
}
}
\author{
Main Developer and Maintainer
Louis, Héraut, \href{mailto:louis.heraut@inrae.fr}{louis.heraut@inrae.fr}*1

Other Developer
Benjamin, Renard, \href{mailto:benjamin.renard@inrae.fr}{benjamin.renard@inrae.fr}*2
Valentin, Mansanarez

Other Contributors
Éric, Sauquet, \href{mailto:eric.sauquet@inrae.fr}{eric.sauquet@inrae.fr}*1
Jean-Philippe, Vidal, \href{mailto:jean-philippe.vidal@inrae.fr}{jean-philippe.vidal@inrae.fr}*1
Nathan, Pellerin, \href{mailto:nathan.pellerin@inrae.fr}{nathan.pellerin@inrae.fr}*1
David, Dorchies. \href{mailto:david.dorchies@inrae.fr}{david.dorchies@inrae.fr}*3

*1 INRAE, UR RiverLy, Villeurbanne, France
*2 INRAE, RECOVER, Aix-Marseille Université, Aix-en-Provence, France
*3 INRAE, UMR G-Eau, Montpellier, France
}
