% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trend.R
\name{get_period}
\alias{get_period}
\title{Period of trend analysis}
\usage{
get_period(data, trendEX, verbose = TRUE)
}
\arguments{
\item{data}{A tibble containing at least the following columns: a date column, a unique numeric column for the values to be analyzed, and a character column for the names of the time series used for grouping.}

\item{trendEX}{A tibble containing the result of the trend analysis}

\item{verbose}{If set to true, function will output the progress of the computation}
}
\value{
A tibble containing the trend analysis results with additional columns for the start and end of the period.
}
\description{
Computes the start and end of the period for a trend analysis based on the available data.
}
\note{
documentation generated by chatGPT
}
