% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trend.R
\name{get_valueExtremes}
\alias{get_valueExtremes}
\title{Extremes}
\usage{
get_valueExtremes(
  dataEX,
  metaEX,
  trendEX,
  take_not_signif_into_account = TRUE,
  period_change = NULL,
  exProb = 0.01
)
}
\description{
Extremes
}
