% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{get_valueExtremes}
\alias{get_valueExtremes}
\title{Extremes}
\usage{
get_valueExtremes(
  list_df2plot,
  Code,
  nPeriod,
  nbVar,
  nCode,
  valueType = "trend",
  colorForce = FALSE,
  minXprob = 0,
  maxXprob = 1
)
}
\description{
Extremes
}
