% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_Xtrend}
\alias{get_Xtrend}
\title{X trend}
\usage{
get_Xtrend(
  var,
  df_data,
  df_meta,
  period,
  hydroPeriod,
  df_flag = NULL,
  yearNA_lim = NULL,
  dayNA_lim = NULL,
  day_to_roll = NULL,
  functM = NULL,
  functM_args = NULL,
  isDateM = FALSE,
  functY = NULL,
  functY_args = NULL,
  isDateY = FALSE,
  functYT_ext = NULL,
  functYT_ext_args = NULL,
  isDateYT_ext = FALSE,
  functYT_sum = NULL,
  functYT_sum_args = NULL,
  df_mod = tibble()
)
}
\description{
X trend
}
