% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{missing_day}
\alias{missing_day}
\title{Missing period over several days}
\usage{
missing_day(
  df_data,
  dayLac_lim = 3,
  hydroYear = "01-01",
  df_mod = NULL,
  verbose = TRUE
)
}
\description{
Missing period over several days
}
