% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trend.R
\name{process_trend}
\alias{process_trend}
\title{process_trend}
\usage{
process_trend(
  dataEX,
  metaEX = NULL,
  MK_level = 0.1,
  timeDep_option = "INDE",
  take_not_signif_into_account = TRUE,
  period_trend = NULL,
  period_change = NULL,
  exProb = 0.01,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{MK_level}{Numeric value between 0 and 1, the level used in the Mann-Kendall statistical test.}

\item{timeDep_option}{Character indicating the time dependency option of the MK test.}

\item{verbose}{Logical, if true will print detailed information.}

\item{...}{other arguments passed to function \code{GeneralMannKendall_WRAP}}

\item{data}{A tibble with at least three columns : date, numeric values, and a character column for names of time series.
The trend analysis will be performed on the numeric column.}
}
\value{
A table which specifies per line the name of the time series the value 'p', the value 'stat' and 'a' the slope of Theil-Sen. If the 'p' value is less than 0.1 and the slope 'a' is positive, the associated time serie shows an increasing linear trend which can be represented by the equation 'Y = a*X + b ' with 'a' the Type I error of 10\% or 90\% confidence.
}
\description{
Applies Mann-Kendall statistical test to extracted
variable in order to determine the existence or not of a trend in
the data.
}
\note{
documentation generated by chatGPT
}
\examples{
# Date
Start = as.Date("1972-01-01")
End = as.Date("2020-12-31")
Date = seq.Date(Start, End, by="day")

# Value to analyse
set.seed(100)
X = seq(1, length(Date))/1e4 + runif(length(Date), -100, 100)
X[as.Date("2000-03-01") <= Date & Date <= as.Date("2000-09-30")] = NA

# Creation of tibble
data = dplyr::tibble(Date=Date, ID="serie A", X=X)

# Extraction
dataEX = process_extraction(data=data,
                            samplePeriod=c("05-01",
                                           "11-30"),
                            funct=max,
                            na.rm=TRUE,
                            period=c(as.Date("1990-01-01"),
                                     as.Date("2020-12-31")),
                            timeStep="year")

trendEX = process_trend(data=dataEX)
trendEX
}
