% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{sourceProcess}
\alias{sourceProcess}
\title{sourceProcess}
\usage{
sourceProcess(path, default = NULL)
}
\arguments{
\item{path}{Path to the script file}

\item{default}{Default process object to use (optional)}
}
\value{
A list of process objects extracted from the script file.
}
\description{
Source and process a script file.
}
\details{
The function sources the script file from the specified path and extracts the process objects defined within the file. It returns a list of process objects, where each object represents a different process defined in the script file.
}
