% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_process.R
\name{sourceProcess}
\alias{sourceProcess}
\title{sourceProcess}
\usage{
sourceProcess(path, default = NULL)
}
\arguments{
\item{path}{A \link[base:character]{character} string as a path to the CARD file.}

\item{default}{The default process loaded by \code{Process_default = sourceProcess(file.path(CARD_path, "__default__.R"))} to load the default parameters of every extraction process. Default \code{NULL}.}
}
\value{
A \code{Process} variable which is a list of general parameters, and sub list for each extraction process parameters.
}
\description{
Allows read CARD with is formatism and convert it to a \code{Process} variable that explain all the step necessary for the extraction of \code{\link[=process_extraction]{process_extraction()}}.
}
\examples{
\dontrun{
Process_default = sourceProcess(path="path/to/CARD/__default__.R")
Process = sourceProcess(path="path/to/CARD/script.R",
                        default=Process_default)
}
}
\seealso{
\itemize{
\item \code{\link[=process_extraction]{process_extraction()}} for extracting variables.
\item \code{\link[=process_trend]{process_trend()}} for performing trend analysis on extracted variables.
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
\item \code{\link[=CARD_extraction]{CARD_extraction()}} for extracting variables using CARD parameterization files.
}
}
